Small fix.
svn path=/trunk/kdeextragear-1/amarok/; revision=327156

+++ src/engine/gst/gstengine.cpp
@@ -398,9 +398,11 @@ GstEngine::stop()  //SLOT
     /* stop the thread */
     gst_element_set_state ( m_thread, GST_STATE_NULL );
 
-    m_transferJob->kill();
-    m_transferJob = 0;
-    
+    if ( m_transferJob ) {
+        m_transferJob->kill();
+        m_transferJob = 0;
+    }
+            
     emit stopped();
 }
